+++ src/versionUtils.ts
@@ -24,13 +24,13 @@ import * as kamelCli from './kamel';
 import { platformString } from './installer';
 import fetch from 'cross-fetch';
 
-export const version: string = '1.3.2'; //need to retrieve this if possible, but have a default
+export const version: string = '1.4.0'; //need to retrieve this if possible, but have a default
 
 /*
-* Can be retrieved using `curl -i https://api.github.com/repos/apache/camel-k/releases/latest` and searching for "last-modified" attribute
-* To be updated when updating the default "version" attribute
-*/
-const LAST_MODIFIED_DATE_OF_DEFAULT_VERSION: string = 'Thu, 25 Mar 2021 09:12:01 GMT';
+ * Can be retrieved using `curl -i https://api.github.com/repos/apache/camel-k/releases/latest | grep last-modified`
+ * To be updated when updating the default "version" attribute
+ */
+const LAST_MODIFIED_DATE_OF_DEFAULT_VERSION: string = 'Tue, 20 Apr 2021 11:25:24 GMT';
 let latestVersionFromOnline: string;
 
 export async function testVersionAvailable(versionToUse: string): Promise<boolean> {